+++ app/src/main/java/com/example/playlistmaker/presentation/PlayerActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.presentation
 
 import android.content.ContentValues
 import android.media.MediaPlayer
@@ -7,7 +7,6 @@ import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
-import android.widget.Button
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
@@ -15,16 +14,15 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.R
+import com.example.playlistmaker.data.PlayerImpl
 import java.text.SimpleDateFormat
 import java.util.Locale
 
 class PlayerActivity : AppCompatActivity() {
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     private var mainThreadHandler: Handler? = null
-    private var mediaPlayer = MediaPlayer()
-    private lateinit var play: ImageButton
-    private lateinit var timer: TextView
-    private lateinit var url: String
+
     companion object {
         private const val STATE_DEFAULT = 0
         private const val STATE_PREPARED = 1
@@ -32,10 +30,24 @@ class PlayerActivity : AppCompatActivity() {
         private const val STATE_PAUSED = 3
     }
 
-    private var playerState = STATE_DEFAULT
+
+    private lateinit var play: ImageButton
+    private lateinit var timer: TextView
+    private lateinit var url: String
+
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
+    var player = PlayerImpl()
+
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
+    var mediaPlayer = player.mediaPlayer
+
+    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
+    private var playerState = player.playerState
+
 
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onCreate(savedInstanceState: Bundle?) {
+
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_player)
         mainThreadHandler = Handler(Looper.getMainLooper())
@@ -50,7 +62,7 @@ class PlayerActivity : AppCompatActivity() {
         val genreValue = findViewById<TextView>(R.id.genreValue)
         val countryValue = findViewById<TextView>(R.id.countryValue)
         play = findViewById(R.id.playPauseButton)
-        timer =  findViewById(R.id.playTime)
+        timer = findViewById(R.id.playTime)
         url = intent.extras?.getString("previewUrl").toString()
 
 
@@ -76,13 +88,20 @@ class PlayerActivity : AppCompatActivity() {
         genreValue.text = intent.extras?.getString("primaryGenreName") ?: ""
         countryValue.text = intent.extras?.getString("country") ?: ""
 
-        preparePlayer()
+        player.preparePlayer(
+            play = play,
+            url = url,
+            timer = timer,
+            mainThreadHandler = mainThreadHandler
+        )
+        Log.v(ContentValues.TAG, "playerState is $playerState")
         play.setOnClickListener {
             Log.v(ContentValues.TAG, "START PLAYING")
-            playbackControl()
+            player.playbackControl(play = play)
+
 
             mainThreadHandler?.post(
-                createUpdateTimerTask()
+                player.createUpdateTimerTask(timer = timer, mainThreadHandler = mainThreadHandler!!)
             )
         }
 
@@ -91,79 +110,24 @@ class PlayerActivity : AppCompatActivity() {
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onPause() {
         super.onPause()
-        pausePlayer()
     }
+
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onDestroy() {
         super.onDestroy()
         mediaPlayer.release()
-        mainThreadHandler?.removeCallbacks(createUpdateTimerTask())
+        mainThreadHandler?.removeCallbacks(
+            player.createUpdateTimerTask(
+                timer = timer,
+                mainThreadHandler = mainThreadHandler!!
+            )
+        )
     }
 
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private fun preparePlayer() {
-        mediaPlayer.setDataSource(url)
-        mediaPlayer.prepareAsync()
-        mediaPlayer.setOnPreparedListener {
-            play.isEnabled = true
-            playerState = STATE_PREPARED
-        }
-        mediaPlayer.setOnCompletionListener {
-            mainThreadHandler?.removeCallbacks(createUpdateTimerTask())
-            playerState = STATE_PREPARED
-            timer.text ="00:00"
-            play.setBackgroundResource(R.drawable.playpausebutton)
 
-        }
-    }
-
-    private fun createUpdateTimerTask(): Runnable {
-        return object : Runnable {
-            @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-            override fun run() {
-                Log.v(ContentValues.TAG, "TIMER TASK")
-                when(playerState) {
-                    STATE_PLAYING -> {
-                        timer.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(mediaPlayer.currentPosition)
-                        mainThreadHandler?.postDelayed(this,500)
-
-                    }
-                    STATE_PREPARED, STATE_PAUSED -> {
-
-                        //timer.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(mediaPlayer.currentPosition)
-                        mainThreadHandler?.removeCallbacks(this)
-                    }
-                }
-                Log.v(ContentValues.TAG, "Time is $mediaPlayer.currentPosition")
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private fun startPlayer() {
-        mediaPlayer.start()
-        play.setBackgroundResource(R.drawable.pausebutton)
-        playerState = STATE_PLAYING
-    }
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private fun pausePlayer() {
-        mediaPlayer.pause()
-        play.setBackgroundResource(R.drawable.playpausebutton)
-        playerState = STATE_PAUSED
-    }
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private fun playbackControl() {
-        when(playerState) {
-            STATE_PLAYING -> {
-                pausePlayer()
-            }
-            STATE_PREPARED, STATE_PAUSED -> {
-                startPlayer()
-            }
-        }
-    }
 }
+
+
+
